Move the dump formats code from lib to cli
