+++ Cargo.toml
@@ -25,12 +25,8 @@ petgraph = "^0.5"
 phf = { version = "^0.8", features = ["macros"] }
 regex = "^1.1"
 serde = { version = "^1.0", features = ["derive"] }
-serde_cbor = "^0.11"
-serde_json = "^1.0"
 termcolor = "^1.0"
 tree-sitter = "^0.16"
-serde_yaml = "^0.8"
-toml = "^0.5"
 
 [dev-dependencies]
 pretty_assertions = "^0.6"
