+++ rust-code-analysis-cli/Cargo.toml
@@ -16,9 +16,13 @@ crossbeam = "^0.7"
 futures = "^0.3"
 globset = "^0.4"
 num_cpus = "^1.13"
+regex = "^1.1"
 rust-code-analysis = { path = "..", version = "0.0"}
 serde = "^1.0"
+serde_cbor = "^0.11"
 serde_json = "^1.0"
+serde_yaml = "^0.8"
+toml = "^0.5"
 walkdir = "^2.2"
 
 [dev-dependencies]
