+++ src/lib.rs
@@ -48,10 +48,6 @@
 extern crate lazy_static;
 #[macro_use]
 extern crate serde;
-extern crate serde_cbor;
-extern crate serde_json;
-extern crate serde_yaml;
-extern crate toml;
 
 #[macro_use]
 mod asttools;
