+++ src/output/mod.rs
@@ -1,8 +1,5 @@
 pub(crate) mod dump;
 pub use dump::*;
 
-pub(crate) mod dump_formats;
-pub use dump_formats::*;
-
 pub(crate) mod dump_metrics;
 pub use dump_metrics::*;
